Fix install and uninstall scripts
* Check if system uses /etc/init.d/.
* Remove unnecessary deletions.

+++ bin/install.sh
@@ -2,8 +2,9 @@
 chmod +x $(pwd)/bin/uninstall.sh
 chmod +x $(pwd)/bin/impress
 sudo ln -s -f $(pwd)/cli.js /bin/impress
-sudo rm -f /etc/init.d/impress
-sudo ln -f $(pwd)/bin/impress /etc/init.d/impress
+if [ -f /etc/init.d/ ]; then
+  sudo ln -f $(pwd)/bin/impress /etc/init.d/impress
+fi
 if [ "$(uname -s)" = 'FreeBSD' ]; then
   ln -f $(pwd)/impress-freebsd.sh /etc/rc.d/impress
   chmod +x /etc/rc.d/impress
@@ -12,7 +13,6 @@ if [ "$(uname -s)" = 'FreeBSD' ]; then
 elif [ -f /etc/debian_version ]; then
   sudo update-rc.d impress defaults
 elif [ -n "$(command -v systemctl)" ]; then
-  sudo rm -f /etc/systemd/system/impress.service
   sudo ln -f $(pwd)/bin/impress.service /etc/systemd/system/impress.service
   sudo systemctl daemon-reload
   sudo systemctl enable impress
